Agent: Move MultiprocessingPluginWrapper out of plugin_loader
